Add workflow for creating JKS keystore as a repo secret
And polish the existing ones.

The idea is to use Github Actions to generate the key pair used for
signing the APKs, and store the resulting key store as secrets in the
repository, available for use by the "build and release" actions.

This way, no one contributor needs to be entrusted with the key
generation, minimizing the opportunities to keep a copy of the app
signing key.

The only way to read repository secrets is to create and run an
Action/workflow where the secrets are printed in an obfuscated manner
(to prevent Github from masking them). This can of course be done by
repo/organization admins, but it is hard to do without leaving any
traces.

In order to create repository secrets from a workflow, a fine-grained
personal access token is required when running the "Create APK signing
secrets" workflow. Since this workflow is only really intended to be run
once, and then disabled (but kept as documentation), I recommend
creating a very short-lived personal access token for use by this
workflow.

If anyone else wants to trigger the workflow to overwrite the app
signing keys, they must have repo/org admin privileges, so that they can
first create a personal access token with permission to write secrets.