Place Git workdirs in private storage
and remove all Git settings related to the workdir location.

The directory containing the workdir is named using the repo ID. This
should allow changing the URL of an existing repo.

I thought for a while about a feature to copy/export the directory to
public storage for troubleshooting purposes, but I came to the
conclusion that it should not be necessary. If the user is worried about
losing local state, they can always export the relevant notebooks as
text files, and re-import them later. If the repo is somehow broken, it
is best to delete it and add a new one, re-linking the notebooks (and
deleting the local copies, if necessary). If there are tenacious bugs,
they should be ironed out on a virtual device with full root access.

This resolves #21.