Set a lower threshold for Git garbage collection
The number refers to the threshold for loose objects in the repo before
Git performs housekeeping tasks in connection with common porcelain
commands such as "push". The default value is 6700.

This is just intended as a first step towards finding a suitable limit.
It should be evaluated by checking regularly how much space can be saved
in a repo by running "git gc" manually. If a considerable amount of space
can be saved after, say, one month of heavy repo use, then this limit
should probably be lowered further.