Git: Further lower the garbage collection threshold
I just noticed that my day-to-day repo could be shrunk from >6 MB to 2.7
MB by running a regular "git gc". This means that more than half of its
content was garbage, which should have been collected.

I also confirmed that a recent clone of the repo is much more nippy in
Orgzly.

Related to issue #27.