Log sync events to major events app log
This enables easier benchmarking of sync times when testing changes to
the sync logic. I am currently looking at alternative sync workflows for
Git repos, and it's hard to compare sync times just by looking.

Why did I remove the argument to AppLogsRepository.getFlow()? Because I
believe it was a case of over-engineering. There is no way to specify
which type of logs you want to view, and until there is, the only
reasonable behavior is to show all logs.