Revert "Upgrade JGit to 6.7.0 and add Proguard keep settings"
This reverts commit e73d9f4041fc5ed61174da23570c932ca79c4157.

If we use any JGit version above 6.2.x (as of this writing, the latest
available version is 6.8.0.202311291450-r), we run into this problem:
https://stackoverflow.com/questions/76549421/jgit-failure-in-cloning-repository.
Naturally, we don't want to raise minSdkVersion to 33.

But JGit 6.2.x depends on a version of the Apache MINA SSHD library
(2.8.0) which does not handle Android well (see
https://github.com/apache/mina-sshd/issues/322). A later version
(2.10.0) added basic Android awareness, but that version is not used by
JGit until in v6.6.x.

We are thus stuck using JGit v5 until the Android support in JGit v6
improves.