Improved handling of remotely deleted files
This resolves #77.

We are adding two new BookSyncStatuses:

- ROOK_NO_LONGER_EXISTS
- BOOK_WITHOUT_LINK_AND_PREVIOUS_ERROR

The first one is pretty obvious, but the second one requires some explanation.

When a remote file is deleted, I have chosen to do two things to
Orgzly's namesake notebook:

1. remove the book's repository link
2. clear the book's "synced to" attribute (which typically holds the
   last seen revision ID of the repository file)

I chose to do this because it then becomes much easier to draw conclusions
about the book's state during subsequent syncing attempts.

One of the things it allows us to do is to NOT automatically link the
notebook to the configured repository and sync it. This was the annoying
previous behavior which we want to get rid of. For background, when
Orgzly sees a local notebook without a repository link, and there is
only one configured repository, it will usually silently link the
notebook to the repository and sync it. This is great when the user has
created a new notebook in Orgzly, but it's not desirable when something
more unusual has happened, such as a file deletion. Instead, we want the
user to be alerted about the deleted file, and then take action
themselves. They probably want to delete the notebook from Orgzly, but
they may also want to link it to a different repo, or back to the same
one, in case the deletion was a mistake.

In other words, the new sync status BOOK_WITHOUT_LINK_AND_PREVIOUS_ERROR
allows us to distinguish between a local book which can be automatically
linked to a repo and synced, and one which should not be touched, but
left in an error state until the user takes action.