Issue #381: Avoid guessing names from content:// URLs
Some DocumentFiles do not work at all with the guessing of file and
directory names from content:// URLs, which I introduced in #312. (I knew
it was sort of wrong, but it seemed to work...)

This solution should not be noticeably slower than the previous one. The
only situation in which we need to walk the file tree from top to bottom
in order get the full repo-relative path from a file URL, is when we are
renaming files. And that is very rare operation which doesn't need to be
fast.

The biggest change here is in the signature of SyncRepo.retrieveBook,
which now takes a URI, instead of a repo-relative path string. This
required poking in quite a few different places, but the tests are
paying off.