Issue #174: Parse book properties contained in preface
This presupposes https://github.com/orgzly-revived/org-java/pull/6.

Linking to notebooks via their "ID" property now works.

I did not touch the handling of the preface, so the book properties are
visible as part of the preface content. This means they can be shown and
edited in Orgzly without changing the current GUI. It feels a bit hacky,
but it could be a first, simple implementation of book properties, which
I suppose could later be improved by adding a "edit book properties"
menu choice, similar to today's "edit preface".

Book properties are stored in a separate DB table, just like note
properties. They are only parsed and stored when loading books from files and
when the preface has been edited.