[WIP] Provide an interface for defining and using note templates
This commit contains some small cleanups which support the idea of using
templates. It includes broader usage of NotePayload as object for
storing template data.

With this commit the user has the option to decide whether to start with
a clean note or with a template. Templates are defined as regular notes
in a book. Templates are identified by a property called
"orgzly_template". Every note that has this property set will be
available as template. There is no dedicated user interface for
maintaining templates, as this can implicitly be done via book/note
views.
Templates only store data, no position. Consequently, you cannot use
templates to create notes as children to a specific other note.
The property value is currently used as title to identify the templates.
In the future this could be used to identify types of templates.

TODO:
- polish user interface for selecting templates (multi-row for better
  distinguishment (title vs property vs summary)
- assign a template to widget instance (using the Plus-Sign can then
  provide the user with a note that already has prefilled data)
- Introduce settings to enable / disable templates as the new template
  implementation has an impact on workflows
- another round of code cleanup and review
- centralize string identifiers like "orgzly_template"
- ...