Issue #128: Add feature to export/import settings to a chosen note
After successful export, the ID property of the target note is stored to
preferences. Currently, this is only used for opening the import or
export "note picker" in the right place, so it's a bit pointless. But I'm
thinking in the future this might allow for automatic export whenever
settings are changed.

The UI code has mostly been copied from ui.refile. If I were better at
Android, I would refactor to reduce the amount of duplicated code. But
this is easier, and I doubt it will create total chaos.