Adding parent title update to setNotesState per PR feedback
The app lets users hold to select multiple notes and then update their states all in one go, which prior to this change would not update the title cookies of parent notes when necessary.

Also no longer checking whether the new state is DONE or TODO when deciding whether to update the title of the parent because all the possible transitions aren't covered and the tryUpdate methods shouldn't make any changes unless the title should actually be updated anyway.