Issue #536, #218: Don't read mtime from file system
This is also a reversion back to the code before my commit 43d88fd.
Reading the mtime from the file system at this point works with some
ContentProviders, but with others, that metadata is not available until
some moments later. Which leaves us with no choice but storing our own
timestamp, and then hoping that it will match the file's metadata upon
the next sync occasion.