Issue #536, #218: Go back to storing mtime as revision string
This is truly a reversal to Neven's code before I started fiddling with these lines.

The "revision" string is what we actually compare when deciding whether
the repo file has changed. A millisecond difference between our value
and the filesystem's value is enough to throw us off. When I changed
this to using the current time, instead of asking the filesystem, the
issues became worse.

I still don't know what to do about ContentProvider's which return a
null mtime at this point, but at least this code is not worse than the
original Orgzly code...