[bug] fixed #131: Note contents not fully shown.
Please see investigation notes in the issue #131. For future reference, it's tricky issue that missing one attribute android:inputType=textMultiLine implicitly causes a slightly different bug of text size limit. If we're building a new use case from RichText, not knowing the connection makes it very difficult to test and debug. A good solution could be to use a subclass MultilineRichText to make this explicit, so engineers can make a better choice.