Issue #332: ShareActivity: Put text payload in note body
instead of the title, and allow the user to input a heading - in the
case when there is only a TEXT payload, and nothing else.

If we are receiving both TEXT and SUBJECT payloads, we will put the
former in the body and the latter as the suggested heading.

If we're receiving a URL accompanied by a subject/title, that will get
turned into a heading with a link, and no body.