RetryTestRule: Do 10 retries by default instead of 3
This will make us more certain that a failing test really has serious
issues.

I have also tried out Barista, and it seemed to work fine, but I
realized that the retry feature was the main thing I wanted, and we
already have a solution to that. Also, Barista hasn't had a proper
release since 2022.