Issue #92: Use WorkManager instead of AlarmManager
WorkManager is simply the right tool for the job, here.

I have used this code for a while and it's working as intended. My only
question right now is how to succinctly explain in the settings that
the specified interval is a minimum repeat interval, and that the actual
interval may be extended by sometimes an hour or more, depending on
Android's battery optimization logic. Personally, I'm fine with this,
but maybe it's not very intuitive nor easy to explain. So I see a risk
that users will be confused and wonder why sync happens after e.g. 45
minutes, instead of 15 minutes, as they specified.