Issue #273: Add "Insert inline timestamp" button
When the user is editing either the title or content of a note, the
"insert inline timestamp" button appears. When the user is not editing,
the button is removed.

After inserting the timestamp, we ensure that the keyboard is shown
again, so that editing can continue. (Not sure is this will be a
nuisance to users with physical keyboards, though.)

The TimestampDialogFragment has been updated to include a
active/inactive checkbox, which is only visible when inserting an inline
timestamp.

The user's last choice of active/inactive is saved to app preferences,
to make it easy to quickly insert multiple similar timestamps.

I have renamed the variable "id" to "originViewId" to clarify its
meaning.